requirements.txt
discord.py
openai
add-voice-bot
uberduck
io
pydub
discord-ext-audiorec
tiktoken